refactor: userApi error 객체 만드는 함수 모둘화 (#137)
