+++ src/utils/errorUtile.js
@@ -1,5 +1,18 @@
 import { Error503 } from '../components/common/error';
 
+export const throwResponseError = async ({ response, defaultMessage }) => {
+	const responseText = await response.text();
+	const parsedResponse = JSON.parse(responseText);
+	const error = new Error(defaultMessage);
+	const message = `${error.message} ${parsedResponse.error}`;
+	const newError = {
+		errorCode: response.status,
+		errorMessage: message,
+	};
+
+	throw newError;
+};
+
 export const errorHendler = (formInstance, formContainerEl, error) => {
 	const { errorCode, errorMessage } = error;
